requirements.txt
pyspark==3.5.0
kafka-python
streamlit
pandas
plotly

fastapi
uvicorn
websockets



fastparquet
